Add DataProvider class - implement getData method Update Delete.php - add id in logic Update Edit.php - change _initAction method logic Update GenericButton.php - add (int) in logic Refactor RestaurantDataProvider - change name - modify getData method logic Update restaurants_restaurant_form.xml - edit dataProvider configuration path Update restaurants_restaurant_formnew.xml - edit dataProvider configuration path Update restaurants_restaurant_index.xml Update restaurants_restaurant_listing.xml - add missing configuration U
