+++ app/code/Stoyanov/Restaurant/Ui/Component/New/DataProvider.php
@@ -1,22 +1,21 @@
 <?php
 declare(strict_types=1);
 
-namespace Stoyanov\Restaurant\Model\Restaurant;
+namespace Stoyanov\Restaurant\Ui\Component\New;
 
 use Magento\Ui\DataProvider\AbstractDataProvider;
 use Stoyanov\Restaurant\Model\ResourceModel\Restaurant\CollectionFactory;
 
 class DataProvider extends AbstractDataProvider
 {
-    protected $loadedData;
-
     public function __construct(
         $name,
         $primaryFieldName,
         $requestFieldName,
         CollectionFactory $collectionFactory,
         array $meta = [],
-        array $data = []
+        array $data = [],
+
     ) {
         $this->collection = $collectionFactory->create();
         parent::__construct($name, $primaryFieldName, $requestFieldName, $meta, $data);
@@ -24,13 +23,6 @@ class DataProvider extends AbstractDataProvider
 
     public function getData()
     {
-        if (isset($this->loadedData)) {
-            return $this->loadedData;
-        }
-        $items = $this->collection->getItems();
-        foreach ($items as $restaurant) {
-            $this->loadedData[$restaurant->getId()] = $restaurant->getData();
-        }
-        return $this->loadedData ?? [];
+        return [];
     }
 }
