+++ app/code/Stoyanov/Restaurant/view/adminhtml/ui_component/restaurants_restaurant_form.xml
@@ -30,7 +30,7 @@
         <settings>
             <submitUrl path="restaurants/restaurant/save"/>
         </settings>
-        <dataProvider class="Stoyanov\Restaurant\Model\Restaurant\DataProvider" name="restaurant_form_data_source">
+        <dataProvider class="Stoyanov\Restaurant\Ui\Component\Edit\DataProvider" name="restaurant_form_data_source">
             <settings>
                 <requestFieldName>id</requestFieldName>
                 <primaryFieldName>id</primaryFieldName>
